agregar documentación, agregar funcionalidad para subir imagenes a supabase
